Use literal option hashes instead of kwargs
Because we mostly only have dynamic lists of options, using kwargs meant
just using splats, and the changes in ruby 2.7 and ruby 3.0 to splats
make it harder to keep a single version of the code working with ruby
2.6 and onwards.

For now, explicit option Hashes are good. And the validation is already
happening in HasOptions, so there's no real downside.