Separate conditions to evaluate and to render
A query builder might be renderable, but not evaluatable. For example,
we want to render filters that don't have a value (so you can set a
value!) but we don't want to evaluate them.

This introduces a new method, `QueryBuilder#render?` that defaults to
evaluating the `if`/`unless` options, and changes `#evaluate?` to
default to `render?`.

This means that:

* If the `if`/`unless` options turn off something, then it won't be
  either rendered or evaluated.
* If something overwrites the render behavior and render? returns false,
  then it will also not be evaluated. (If you need this to change, you
  can simply change the evaluate behavior.)
* If something turns off evaluation, you might still render it (i.e.
  non-database columns, or filters without a value.)