Improve the API of EmptyState
title and subtitle are too generic. This changes subtitle for
explanation, which hopefully provides a better idea of what to
put in it, and adds a third attribute, `update_filters`, to show
when filters are too restrictive. (i.e. when the query returns
an empty set and filters are applied.)