Refactor DataTable.build
Now instead of forcing the input to be a Query class, it can be any
object that implements #to_data_table and returns a Hash of data table
arguments.

In turn, Query implements .to_data_table to satisfy this API.

This allows you building a Query in any way you want, and then passing
it to `DataTable.build`.