Allow Queries and DataTables to be renderable
Queries are only renderable as a proxy for DataTables, which are the
ones doing the rendering. But since people define all behavior on the
Query object and just use DataTable as a view-level wrapper, this seems
like a reasonable approach—considering that with the current version
implementations end up defining partial options on the query object
anyway.